Refactoring scripts to start testing
