--- .htaccess
@@ -1,3 +1,0 @@
-RewriteEngine on
-RewriteRule .* /public/$0 [L]
-DirectoryIndex index.php
