+++ src/GraphQl/Resolvers/CategoryResolver.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\GraphQL\Resolvers;
+
+class CategoryResolver
+{
+    public static function findAll(): array
+    {
+        // Exemplo de retorno estático para teste
+        return ['Electronics', 'Clothing', 'Books'];
+    }
+}
