+++ src/GraphQl/Resolvers/ProductResolver.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\GraphQL\Resolvers;
+
+class ProductResolver
+{
+    public static function findAll(): array
+    {
+        // Exemplo de retorno estático para teste
+        return ['Electronics', 'Clothing', 'Books'];
+    }
+}
